#231: Password Reset Spam Prevention
Don't allow user to reset their password if the last password reset was done by the system. They should already have an email with their new password. This will prevent users from being spammed with 100s of emails.

# Overview

- Add statement to check for `passwordReset` field
- Refactored `gmailAuth` to propagate errors up the stack
- Converted clients of `gmailAuth` to use callbacks and handle errors
- Reset system message on admin changing password

## Refactoring `gmailAuth`

`gmailAuth` fails silently if there is an error sending an email.
This is bad because the `resetPassword` route would change the user's
password without notifying the user.

`gmailAuth` was refactored to propagate errors to clients
asynchronously via callbacks.

## Logic changes to `resetPassword`

- Previously: Find user, create new password, *store in DB, send email*
- Now: Find user, create new password, *send email, store in DB*

This means if the email fails the send, we will not change the user's
password in our DB. This is what we want.

# Testing Done

- Invalid email
- No credentials in backend
- Successful test